Moving src/stan/services/mcmc/run_markov_chain.hpp to src/stan/services/sample/generate_transitions.hpp
