+++ make/tests
@@ -143,10 +143,10 @@ src/test/unit/optimization/lbfgs_update_test.cpp: src/test/test-models/good/opti
 
 src/test/unit/services/init/command_init_test.cpp: src/test/test-models/good/services/test_lp.hpp
 src/test/unit/services/io/write_iteration_test.cpp: src/test/test-models/good/services/test_lp.hpp
-src/test/unit/services/mcmc/run_markov_chain_test.cpp: src/test/test-models/good/services/test_lp.hpp
 src/test/unit/services/mcmc/sample_test.cpp: src/test/test-models/good/services/test_lp.hpp
 src/test/unit/services/mcmc/warmup_test.cpp: src/test/test-models/good/services/test_lp.hpp
 src/test/unit/services/optimize/do_bfgs_optimize_test.cpp: src/test/test-models/good/optimization/rosenbrock.hpp
+src/test/unit/services/sample/generate_transitions_test.cpp: src/test/test-models/good/services/test_lp.hpp
 src/test/unit/services/sample/mcmc_writer_test.cpp: src/test/test-models/good/io_example.hpp
 
 src/test/unit/lang/reject/reject_mult_args_test.cpp: src/test/test-models/good/lang/reject_mult_args.hpp
